Change default math lib realisation to own realisation

+++ src/main/java/ru/vsu/cs/finaltaskcg/render_engine/ZBuffer.java
@@ -1,14 +1,14 @@
 package ru.vsu.cs.finaltaskcg.render_engine;
 
 public class ZBuffer {
-    private float[][] buffer;
+    private double[][] buffer;
     private int width;
     private int height;
 
     public ZBuffer(int width, int height) {
         this.width = width;
         this.height = height;
-        this.buffer = new float[height][width];
+        this.buffer = new double[height][width];
         clear();
     }
 
@@ -20,7 +20,7 @@ public class ZBuffer {
         }
     }
 
-    public boolean testAndSet(int x, int y, float z) {
+    public boolean testAndSet(int x, int y, double z) {
         if (x < 0 || x >= width || y < 0 || y >= height) {
             return false;
         }
